backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
redis==5.0.1
pdfkit==1.0.0
openpyxl==3.1.2